submodule rename und filled __init__.py
