+++ src/napari_time_series_plotter/__init__.py
@@ -1,0 +1,3 @@
+from dock_widget import *
+from widgets import *
+from utils import *
