+++ src/napari_time_series_plotter/utils.py
